requirements.txt
amqp
asgiref
async-timeout
billiard
blinker
celery
click
Django==5.1
django-cors-headers
django-redis
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
drf-yasg
psycopg2-binary
python-dotenv
redis
sqlparse
pillow
gunicorn
boto3
django-storages
# backports.zoneinfo
dj-database-url
pandas
openpyxl
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
drf_spectacular
sentry-sdk[django]
django-silk